requirements.txt
Django==4.1.3
psycopg2==2.9.5
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
